+++ User/Boards/bsp_pwm.h
@@ -2,6 +2,8 @@
 
 
 /* Includes ------------------------------------------------------------------*/
+#include "user_math.h"
+
 /* Exported constants --------------------------------------------------------*/
 /* Exported macro ------------------------------------------------------------*/
 /* Exported types ------------------------------------------------------------*/
@@ -10,6 +12,6 @@ typedef enum {
 } BSP_PWM_Channel_t;
 
 /* Exported functions prototypes ---------------------------------------------*/
-int BSP_PWM_Start(BSP_PWM_Channel_t ch);
-int BSP_PWM_Set(BSP_PWM_Channel_t ch, float duty_cycle);
-int BSP_PWM_Stop(BSP_PWM_Channel_t ch);
+int8_t BSP_PWM_Start(BSP_PWM_Channel_t ch);
+int8_t BSP_PWM_Set(BSP_PWM_Channel_t ch, float32_t duty_cycle);
+int8_t BSP_PWM_Stop(BSP_PWM_Channel_t ch);
